Made Landing Page redux compliant, fixed event creating not appearing on Landing Page, TODO: fix on other pages
