+++ client/src/components/activityFinding/ActivityFinder.jsx
@@ -3,7 +3,7 @@ import CurrentActivityContainer from '../../containers/CurrentActivityContainer.
 import FilterBarContainer from '../../containers/FilterBarContainer.jsx';
 // import SearchBar from './SearchBar.jsx';
 import ActivitiesListContainer from '../../containers/ActivityListContainer.jsx';
-import ActivityCreatorContainer from '../../containers/ActivityCreatorContainer.jsx'
+import ActivityCreatorContainer from '../../containers/ActivityCreatorContainer.jsx';
 import FooterBar from '../navigating/FooterBar.jsx';
 
 const ActivityFinder = ({ createActivity }) => {
