+++ client/src/components/landingPage/LandingPage.jsx
@@ -1,25 +1,26 @@
 import React, { Component } from 'react';
-import Flickity from 'flickity';
-import BottomIcon from '../../assets/BottomIcon.png'
+import LandingPageSplash from './LandingPageSplash.jsx';
+import FeaturedActivities from './featuredActivities.jsx';
+import FooterBar from '../navigating/FooterBar.jsx';
+import ActivityCreatorContainer from '../../containers/ActivityCreatorContainer.jsx';
 
-class LandingPage extends Component {
-  componentDidMount(){
-    let elem = document.querySelector('.main-carousel');
-    let flckty = new Flickity(elem, {
-      fullscreen: true,
-    })
+class LandingPageContainer extends Component {
+  constructor(props){
+    super(props);
   }
   render() {
     return (
-      <div className="welcomeScreen grid-full">
-        <div className="landingPage relative">
-            <img className="featuredImage" src="http://ichef.bbci.co.uk/wwfeatures/wm/live/1280_640/images/live/p0/51/v6/p051v6vn.jpg"/>
-            <img className="bottomIcon absolute" src={BottomIcon}/>
+      <div className="grid-full">
+        <div className="grid-root landingPageContainer">
+          <LandingPageSplash />
+          <FeaturedActivities/>
         </div>
+        <FooterBar className='grid-full'/>
+        <div>{this.props.createActivity === 'creating' ? <ActivityCreatorContainer /> : null}</div>
       </div>
     );
   }
 
 }
 
-export default LandingPage;
+export default LandingPageContainer;
